+++ include/LIEF/ELF/Segment.hpp
@@ -27,6 +27,8 @@
 #include "LIEF/iterators.hpp"
 #include "LIEF/span.hpp"
 
+#include "LIEF/ELF/Header.hpp"
+
 #include "LIEF/ELF/enums.hpp"
 
 namespace LIEF {
@@ -56,12 +58,16 @@ class LIEF_API Segment : public Object {
   using it_const_sections = const_ref_iterator<const sections_t&>;
 
   static constexpr uint64_t PT_BIT = 33;
+  static constexpr uint64_t PT_OS_BIT = 53;
   static constexpr uint64_t PT_MASK = (uint64_t(1) << PT_BIT) - 1;
 
   static constexpr uint64_t PT_ARM      = uint64_t(1) << PT_BIT;
   static constexpr uint64_t PT_AARCH64  = uint64_t(2) << PT_BIT;
   static constexpr uint64_t PT_MIPS     = uint64_t(3) << PT_BIT;
   static constexpr uint64_t PT_RISCV    = uint64_t(4) << PT_BIT;
+  static constexpr uint64_t PT_IA_64    = uint64_t(5) << PT_BIT;
+
+  static constexpr uint64_t PT_HPUX     = uint64_t(1) << PT_OS_BIT;
 
   enum class TYPE : uint64_t {
     UNKNOWN       = uint64_t(-1),
@@ -79,6 +85,7 @@ class LIEF_API Segment : public Object {
     GNU_STACK     = 0x6474e551, /**< Indicates stack executability. */
     GNU_PROPERTY  = 0x6474e553, /**< GNU property */
     GNU_RELRO     = 0x6474e552, /**< Read-only after relocation. */
+    PAX_FLAGS     = 0x65041580,
 
     ARM_ARCHEXT   = 0x70000000 | PT_ARM, /**< Platform architecture compatibility info */
     ARM_EXIDX     = 0x70000001 | PT_ARM,
@@ -91,6 +98,26 @@ class LIEF_API Segment : public Object {
     MIPS_ABIFLAGS = 0x70000003 | PT_MIPS,  /**< Abiflags segment. */
 
     RISCV_ATTRIBUTES = 0x70000003 | PT_RISCV,
+
+    IA_64_EXT         = (0x70000000 + 0x0) | PT_IA_64,
+    IA_64_UNWIND      = (0x70000000 + 0x1) | PT_IA_64,
+
+    HP_TLS            = (0x60000000 + 0x0)  | PT_HPUX,
+    HP_CORE_NONE      = (0x60000000 + 0x1)  | PT_HPUX,
+    HP_CORE_VERSION   = (0x60000000 + 0x2)  | PT_HPUX,
+    HP_CORE_KERNEL    = (0x60000000 + 0x3)  | PT_HPUX,
+    HP_CORE_COMM      = (0x60000000 + 0x4)  | PT_HPUX,
+    HP_CORE_PROC      = (0x60000000 + 0x5)  | PT_HPUX,
+    HP_CORE_LOADABLE  = (0x60000000 + 0x6)  | PT_HPUX,
+    HP_CORE_STACK     = (0x60000000 + 0x7)  | PT_HPUX,
+    HP_CORE_SHM       = (0x60000000 + 0x8)  | PT_HPUX,
+    HP_CORE_MMF       = (0x60000000 + 0x9)  | PT_HPUX,
+    HP_PARALLEL       = (0x60000000 + 0x10) | PT_HPUX,
+    HP_FASTBIND       = (0x60000000 + 0x11) | PT_HPUX,
+    HP_OPT_ANNOT      = (0x60000000 + 0x12) | PT_HPUX,
+    HP_HSL_ANNOT      = (0x60000000 + 0x13) | PT_HPUX,
+    HP_STACK          = (0x60000000 + 0x14) | PT_HPUX,
+    HP_CORE_UTSNAME   = (0x60000000 + 0x15) | PT_HPUX,
   };
 
   enum class FLAGS {
@@ -100,7 +127,7 @@ class LIEF_API Segment : public Object {
     R    = 4,
   };
 
-  static TYPE type_from(uint64_t value, ARCH arch);
+  static TYPE type_from(uint64_t value, ARCH arch, Header::OS_ABI os);
   static uint64_t to_value(TYPE type) {
     return static_cast<uint64_t>(type) & PT_MASK;
   }
@@ -269,7 +296,8 @@ class LIEF_API Segment : public Object {
 
   private:
   template<class T>
-  LIEF_LOCAL Segment(const T& header, ARCH arch = ARCH::NONE);
+  LIEF_LOCAL Segment(const T& header, ARCH arch = ARCH::NONE,
+                     Header::OS_ABI os = Header::OS_ABI::SYSTEMV);
 
   LIEF_LOCAL uint64_t handler_size() const;
   LIEF_LOCAL span<uint8_t> writable_content();